[LPM] Speculative fix for Mac builds.
On Mac builds, it looks like calling WriteNum on a size_t was causing
the compiler to select the templated version of WriteNum rather than the
uint64_t version. Hopefully it won't do the opposite (and break on calls
to WriteNum(uint64_t) because of this fix.

Bug: 808155, 769578

+++ testing/libfuzzer/proto/skia_image_filter_proto_converter.cc
@@ -764,7 +764,7 @@ void Converter::WriteNum(const T num) {
     output_.push_back(num_arr[idx]);
 }
 
-void Converter::WriteNum(const uint64_t num) {
+void Converter::WriteNum(const size_t num) {
   CHECK_LE(num, UINT32_MAX);
   WriteNum(static_cast<uint32_t>(num));
 }
